+++ incl/relationships/deleteGJFriendRequests.php
@@ -1,18 +1,15 @@
 <?php
 chdir(dirname(__FILE__));
-require "../lib/connection.php";
+include "../lib/connection.php";
 require_once "../lib/GJPCheck.php";
 require_once "../lib/exploitPatch.php";
+require_once "../lib/mainLib.php";
+$gs = new mainLib();
 if(empty($_POST["targetAccountID"])){
 	exit("-1");
 }
 $accountID = GJPCheck::getAccountIDOrDie();
 $targetAccountID = ExploitPatch::remove($_POST["targetAccountID"]);
-//REMOVING THE REQUEST
-if(!empty($_POST["isSender"]) AND $_POST["isSender"] == 1){
-		$query = $db->prepare("DELETE from friendreqs WHERE accountID=:accountID AND toAccountID=:targetAccountID LIMIT 1");
-}else{
-		$query = $db->prepare("DELETE from friendreqs WHERE toAccountID=:accountID AND accountID=:targetAccountID LIMIT 1");
-}
-$query->execute([':accountID' => $accountID, ':targetAccountID' => $targetAccountID]);
+$query = $db->prepare("DELETE from friendreqs WHERE (accountID = :accountID AND toAccountID = :targetAccountID) OR (toAccountID = :accountID AND accountID = :targetAccountID) LIMIT 1");
+if($query->execute([':accountID' => $accountID, ':targetAccountID' => $targetAccountID])) $gs->logAction($accountID, 30, $targetAccountID);
 echo "1";
